refactor: extract games_helpers.py from games.py
Move 4 query helper functions (_get_owned_game, _count_active_games,
_get_turn_count, _get_max_turn_number) to games_helpers.py.

Follows admin decomposition pattern (PR #210): shared helpers extracted
first, then concern modules.  games.py: 1610 → 1551 lines (-59 net).
